Move runtime facade below the driver
